+++ AffiseAttributionLib/Classes/modules/attribution/AffiseAttributionModule.swift
@@ -46,28 +46,3 @@ public class AffiseAttributionModule :  NSObject,  AffiseAttributionModuleApi {
         return api?.moduleManager.getModulesNames() ?? []
     }
 }
-
-
-extension AffiseAttributionModule {
-    
-    @available(*, deprecated, message: "Method moved to Affise.Module.Link", renamed: "Link.resolve")
-    @objc
-    public func linkResolve(_ url: String, _ callback: @escaping AffiseLinkCallback) {
-        Affise.Module.Link.resolve(url, callback)
-    }
-    
-    @available(*, deprecated, message: "Method moved to Affise.Module.Subscription", renamed: "Subscription.hasModule")
-    public func hasSubscriptionModule() -> Bool {
-        return Affise.Module.Subscription.hasModule()
-    }
-    
-    @available(*, deprecated, message: "Method moved to Affise.Module.Subscription", renamed: "Subscription.fetchProducts")
-    public func fetchProducts(_ productsIds: [String], _ callback: @escaping AffiseResultCallback<AffiseProductsResult>) {
-        Affise.Module.Subscription.fetchProducts(productsIds, callback)
-    }
-    
-    @available(*, deprecated, message: "Method moved to Affise.Module.Subscription", renamed: "Subscription.purchase")
-    public func purchase(_ product: AffiseProduct, _ type: AffiseProductType?, _ callback: @escaping AffiseResultCallback<AffisePurchasedInfo>) {
-        Affise.Module.Subscription.purchase(product, type, callback)
-    }
-}